pyproject.toml
[project]
name = "interview"
version = "0.1.0"
description = ""
authors = [
    {name = "Justk1dd1ng",email = "tantiarella@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core",
    "langchain-openai",
    "langchain-anthropic",
    "python-dotenv",
    "requests",
    "loguru",
    "langgraph",
    "pydantic>=2.0",
    "httpx"

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
